% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmtree.R
\name{fit_lc}
\alias{fit_lc}
\title{Fit latent-class multinomial models}
\usage{
fit_lc(dataset, data, model, id = NULL, condition = NULL)
}
\arguments{
\item{dataset}{Character.}

\item{data}{A \code{data.frame} containing the data.}

\item{model}{A model definition, typically the path to an \code{.eqn} file.}

\item{id}{Character. Name of the column that contains the subject identifier.
If not specified, it is assumed that each row represents observations from one participant.}

\item{condition}{Character. Name of the column specifying a between-subjects factor.
If not specified, no between-subjects comparisons are performed.}
}
\description{
Does a lot of nice stuff
}
\keyword{internal}
