% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mptinr.R
\name{get_pb_output}
\alias{get_pb_output}
\title{Parametric Bootstrap for MPT}
\usage{
get_pb_output(i, fit_mptinr, data, model_file, col_freq, MPTINR_OPTIONS)
}
\description{
Helper function for creating parametric-bootstrap confidence intervals.
}
\keyword{internal}
