% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_row.R
\name{make_results_row}
\alias{make_results_row}
\title{Create}
\usage{
make_results_row(model, dataset, pooling, package, method, data, id,
  condition, core = NULL)
}
\arguments{
\item{model}{Character.}

\item{dataset}{Character.}

\item{pooling}{Character.}

\item{package}{Character.}

\item{method}{Character.}

\item{data}{A \code{data.frame}.}

\item{id}{Character.}
}
\description{
Internal function, creates container for results
}
\keyword{internal}
