% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{mpt_options}
\alias{mpt_options}
\title{Options Settings for MPT Comparison}
\usage{
mpt_options(...)
}
\description{
Set and examine a variety of \emph{options} which affect the way MPT models
are estimated.
}
\examples{
# Examine options:
mpt_options()

# Set number of MCMC chains to 20:
mpt_options(n.chain = 20)
mpt_options()

}
