% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot.multiverseMPT}
\alias{plot.multiverseMPT}
\title{Plot multiverseMPT}
\usage{
\method{plot}{multiverseMPT}(x, which = "est", save = FALSE,
  write.csv = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{multiverseMPT}.}

\item{which}{Character. Which information do you want to be plotted? Possible
values are \code{"est"} for parameter estimates, \code{"gof1"} for goodness-of-fit statistics,
\code{"test_between"} for ... and \code{"gof2"} for ...}

\item{save}{Logical.}

\item{write.csv}{Logical.}

\item{...}{Further arguments, currently ignored.}
}
\description{
Plot the output from a multiverse MPT analysis.
}
