% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{write_check_results}
\alias{write_check_results}
\title{Write check_results}
\usage{
write_check_results(DATA_FILE, results)
}
\arguments{
\item{DATA_FILE}{Character. File name to use.}

\item{results}{An object of class multiverseMPT.}
}
\description{
Helper function to write the output from check_results() to a file.
}
