% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{write_results}
\alias{write_results}
\title{Write Results of Multiverse Analysis to csv-Files}
\usage{
write_results(results, path = "MPTmultiverse_")
}
\arguments{
\item{results}{An object of class multiverseMPT.}

\item{path}{a path where to save the files (e.g., \code{"C:/results/modelX_dataY_"})}
}
\description{
Exports the results to csv format.
}
