% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_results}
\alias{check_results}
\title{Check results from a multiverse analysis}
\usage{
check_results(results)
}
\arguments{
\item{results}{An object of class multiverseMPT.}
}
\description{
This is a helper function to see if the model estimation worked as intended.
}
