% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mpt.R
\name{fit_mpt}
\alias{fit_mpt}
\title{Main function}
\usage{
fit_mpt(method, dataset, data, model, id = NULL, condition = NULL)
}
\arguments{
\item{method}{Character. A vector specifying which analysis approaches should be performed.}

\item{dataset}{Character.}

\item{data}{A \code{data.frame} containing the data.}

\item{model}{A model definition, typically the path to an \code{.eqn} file.}

\item{id}{Character. Name of the column that contains the subject identifier.
If not specified, it is assumed that each row represents observations from one participant.}

\item{condition}{Character. Name of the column specifying a between-subjects factor.
If not specified, no between-subjects comparisons are performed.}
}
\description{
Does a lot of nice stuff
}
