% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mpt.R
\name{fit_mpt}
\alias{fit_mpt}
\title{Main function}
\usage{
fit_mpt(method, dataset, data, model, id = "id",
  condition = "condition", core = NULL)
}
\arguments{
\item{method}{Character. A vector specifying which analysis approaches should be performed (see Description below).}

\item{dataset}{Character. The name of the dataset to be analyzed.}

\item{data}{A \code{data.frame} containing the data.}

\item{model}{A model definition, typically the path to an \code{.eqn} file.}

\item{id}{Character. Name of the column that contains the subject identifier.
If not specified, it is assumed that each row represents observations from one participant.}

\item{condition}{Character. Name of the column specifying a between-subjects factor.
If not specified, no between-subjects comparisons are performed.}

\item{core}{character vector defining the core parameters of interest, e.g.,
\code{core=c("Dn", "Do")}. All other parameters are treated as auxiliary parameters.}
}
\description{
Does a lot of nice stuff
}
\details{
Maximum-likelihood estimation with MPTinR:
\itemize{
  \item{\code{"asymptotic_complete"}: }{Asymptotic ML theory, complete pooling}
  \item{\code{"asymptotic_no"}: }{ Asymptotitc ML theory, no pooling}
  \item{\code{"pb_no"}: }{Parametric bootstrap, no pooling}
  \item{\code{"npb_no"}: }{Nonparametric bootstrap, no pooling}
}
 
Bayesian estimation with TreeBUGS
\itemize{
  \item{\code{"simple"}: }{Bayesian estimation, no pooling (C++, \link[TreeBUGS]{simpleMPT})}
  \item{\code{"simple_pooling"}: }{Bayesian estimation, complete pooling (C++, \link[TreeBUGS]{simpleMPT})}
  \item{\code{"trait"}: }{latent-trait model, partial pooling (JAGS, \link[TreeBUGS]{traitMPT})}
  \item{\code{"trait_uncorrelated"}: }{latent-trait model without correlation parameters, partial pooling (JAGS, \link[TreeBUGS]{traitMPT})}
  \item{\code{"beta"}: }{beta-MPT model, partial pooling (JAGS, \link[TreeBUGS]{betaMPT})}
  \item{\code{"betacpp"}: }{beta-MPT model, partial pooling (C++, \link[TreeBUGS]{betaMPTcpp})}
}
}
\examples{
examples/examples.fit_mpt.R

}
