% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt_options.R
\name{mpt_options}
\alias{mpt_options}
\title{Options Settings for MPT Comparison}
\usage{
mpt_options(...)
}
\arguments{
\item{...}{Named parameters to set. Possible values are:
\describe{
  \item{\strong{bootstrap_samples}}{Numeric. The number of bootstrap samples to be drawn for the calculation parametric bootstrap confidence intervals.}
  \item{\strong{n.optim}}{Numeric. The number of optimization runs for the models estimated with maximum-likelihood methods.}
  \item{\strong{n.chain}}{Numeric. The number of MCMC chains for the Bayesian models.}
  \item{\strong{n.adapt}}{Numeric. The number of iterations for adaptation.}
  \item{\strong{n.burnin}}{Numeric. The number of burn-in/warm-up iterations.}
  \item{\strong{n.iter}}{Numeric. The total number of iterations to be drawn \emph{after} adaptation (including burnin).}
  \item{\strong{n.thin}}{Numeric. Thinning interval.}
  \item{\strong{Rhat_max}}{Numeric. The maximum rhat.}
  \item{\strong{Neff_min}}{Numeric. The minimum number of effective samples you are willing to accept.}
  \item{\strong{extend_max}}{Numeric.}
  \item{\strong{n.PPP}}{Numeric. The number of posterior predictive samples drawn for the calculation of fit statistics T_1 and T_2.}
  \item{\strong{n.CPU}}{Numeric. The number of CPU cores to use. Defaults to the number of available cores on your machine.}
  \item{\strong{ci_size}}{Numeric.}
  \item{\strong{max_ci_indiv}}{Numeric.}
  \item{\strong{silent}}{Logical.}
  \item{\strong{save_models}}{Logical.}
}}
}
\description{
Set and examine a variety of \emph{options} which affect the way MPT models
are estimated.
}
\examples{
# Examine options:
mpt_options()

# Set number of MCMC chains to 20:
mpt_options(n.chain = 20)
mpt_options()

}
