% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiverseMPT.R
\name{plot.multiverseMPT}
\alias{plot.multiverseMPT}
\title{Plot multiverseMPT}
\usage{
\method{plot}{multiverseMPT}(x, which = "est", save = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{multiverseMPT}.}

\item{which}{Character. Which information should be plotted? Possible
values are
\code{"est"} for parameter estimates,
\code{"test_between"} for between-subjects comparisions,
\code{"gof1"} for overall goodness-of-fit statistics, and
\code{"gof2"} for group-wise goodness-of-fit statistics.}

\item{save}{Logical. Indicates whether the plot should also be saved as a .pdf file.}

\item{...}{ignored.}
}
\description{
Plot the results from a multiverse MPT analysis.
}
