% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot.multiverseMPT}
\alias{plot.multiverseMPT}
\title{Plot multiverseMPT}
\usage{
\method{plot}{multiverseMPT}(x, which = "est", save = FALSE,
  write.csv = FALSE, ...)
}
\arguments{
\item{An}{object of class \code{multiverseMPT}.}
}
\description{
A nice description of the plot method
}
